requirements.txt
requests~=2.32.3
twitchAPI~=4.3.1
discord.py~=2.4.0